Options now carry forward across releases. This commit message is also a follow up to the previous commit. I'm no longer embarrassed by the code behind the options page. Now it's organized, simpler, and better because:
1. Modes are separated from options. My original thinking was that they should be tightly integrated
because most options only make sense in the context of a particular mode. This led to clutter in the modes
code and rigidity in the options code.

2. Options are stored in a sensical way. Previously I stored deeply nested objects in local storage and
performed complex processing to extract parts of interest. Now, I simply request the parts I want. This
had the added benefit of making it easy to communicate changes between the options page(s) and storage.
The redux state is an IDENTICAL COPY of what's in local storage.

3. Previously storage was reset on every update. Now storage is upgraded intelligently upon update. This
code isn't complicated, but it has many corner cases I'll have to write unit tests to verify in the future.

4. Many, many bug fixes

5. Importing and exporting should now be relatively easy to implement and make it into the next commit.